azure-costs/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
PyJWT==2.8.0
cryptography==41.0.7
gunicorn==21.2.0
python-json-logger==2.0.7
requests==2.31.0
azure-mgmt-costmanagement==4.0.1
azure-identity==1.15.0
whitenoise==6.6.0
drf-yasg==1.21.7